Add header file for logger
